Use a versioned S3 bucket with lifecycle for backups
My current backup script is a script I run manually on the last Sunday
of every month, which rsyncs a bunch of stuff to innsmouth, and then a
second script I run on innsmouth to push that into Glacier.  For
various reasons, it's not really robust enough to run unsupervised, so
I can't automate it.

To allow automated backups, I've rewritten my script to use
[duplicity][].  This is great, it can be automated, and it allows for
incremental backups (and so more frequent backups).  But it needs to
be able to read metadata stored on the remote, so it can't use Glacier
as primary storage.

S3 is far more expensive than Glacier, so to trade-off between ease of
access and price, after 60 days backups move into "infrequently
access" storage, and then after a further 120 days into Glacier.  The
backup script only cares about the most recent full backup (and
incremental backups since then), so as long as a full backup is taken
at least once every 180 days, this works out just fine.

[duplicity]: http://duplicity.nongnu.org/